feat: Spring Boot backend with JWT authentication and item management
- JWT Authentication & Security configuration
- User Management (CRUD operations)
- Item Management (Lost/Found items CRUD)
- Item Controller with search and filter
- Item Service with business logic
- Item Repository with JPA
- Item Entity with relationships
- Security configuration updates
- RESTful API endpoints

+++ src/main/java/SampleJWT/auth/entity/User.java
@@ -15,7 +15,7 @@ public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private int id;
+    private Long id;
 
     @Column(name = "firstname", nullable = false)
     private String firstname;
@@ -47,8 +47,8 @@ public class User implements UserDetails {
     }
 
     // 🔹 Getters and Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
 
     public String getFirstname() { return firstname; }
     public void setFirstname(String firstname) { this.firstname = firstname; }
